+++ lib/util/gestures/gestures.dart
@@ -1,2 +1,0 @@
-export 'package:bonfire/util/gestures/drag_gesture.dart';
-export 'package:bonfire/util/gestures/tap_gesture.dart';