+++ lib/util/mixins/pointer_detector.dart
@@ -1,5 +1,6 @@
-import 'package:bonfire/util/gestures/drag_gesture.dart';
-import 'package:bonfire/util/gestures/tap_gesture.dart';
+import 'package:bonfire/gestures/drag_gesture.dart';
+import 'package:bonfire/gestures/mouse_gesture.dart';
+import 'package:bonfire/gestures/tap_gesture.dart';
 import 'package:flutter/gestures.dart';
 
 mixin PointerDetector {
@@ -7,6 +8,8 @@ mixin PointerDetector {
   void onPointerMove(PointerMoveEvent event) {}
   void onPointerUp(PointerUpEvent event) {}
   void onPointerCancel(PointerCancelEvent event) {}
+  void onPointerHover(PointerHoverEvent event) {}
+  void onPointerSignal(PointerSignalEvent event) {}
 }
 
 abstract class PointerDetectorHandler {
@@ -14,10 +17,14 @@ abstract class PointerDetectorHandler {
   void handlerPointerMove(PointerMoveEvent event) {}
   void handlerPointerUp(PointerUpEvent event) {}
   void handlerPointerCancel(PointerCancelEvent event) {}
+  void handlerPointerHover(PointerHoverEvent event) {}
+  void handlerPointerSignal(PointerSignalEvent event) {}
 
   bool hasGesture() {
     if (this is DragGesture && (this as DragGesture).enableDrag) return true;
     if (this is TapGesture && (this as TapGesture).enableTab) return true;
+    if (this is MouseGesture && (this as MouseGesture).enableMouseGesture)
+      return true;
     return false;
   }
 }